Refactor ItemExpression to use less boilerplate
Instead of a bunch of individual fields for each property of
the item to be checked, there is now a set of ItemMatchers,
where ItemExpression.matches(ItemStack) will return false if
any one of the ItemMatchers return false.

This design is much more extendable, since it avoids the
boolean hell the previous design suffered from.

One downside of this design is that the .yaml config format
doesn't at all reflect the API, and making the config format
reflect the API would make the config format nearly unusable.

Another downside is that stages of matching are not
automatically tested on every single item, however this is
more of an advantage because not every stage possible is ran
on every item.